Day 14 - not quite sure what to make of that
On the one hand, it's pretty clever from a design standpoint and needs a
lot of creativity both to code the generation and the solution! On the
other, the lack of ability to test the solution isn't fun, and it's a
bit vague as to what constitutes the answer.

Also some slight perf boosts to older days, and a failed attempt at getting the vs benchmarks going